created packing list controller, pulled deserialized object and assigned values to properties in view model, and reset all of the models in the domain model because hour data was returning null.

+++ GCFinal.Domain/Models/Day.cs
@@ -16,9 +16,9 @@ namespace GCFinal.Domain.Models
         public double totalprecip_mm { get; set; }
         public double totalprecip_in { get; set; }
         public double avgvis_km { get; set; }
-        public double avgvis_miles { get; set; }
-        public double avghumidity { get; set; }
+        public int avgvis_miles { get; set; }
+        public int avghumidity { get; set; }
         public Condition condition { get; set; }
-        public double uv { get; set; }
+        public int uv { get; set; }
     }
 }
